[velox] refactor(functions): Move Presto decimal structs to header
GPU SFI needs to compile Presto decimal simple-function `call()` bodies from device TUs. Nine structs (`DecimalPlusFunction`, `DecimalMinusFunction`, and the other arithmetic/rounding helpers) lived in an anonymous namespace in `DecimalFunctions.cpp`, so no other translation unit could `#include` them.

Move the structs verbatim into `DecimalMathFunctions.h` under `facebook::velox::functions`. Keep registration helpers and `registerDecimal*()` in `DecimalFunctions.cpp`. No logic or registration API changes.

Issue: https://github.com/facebookincubator/velox/issues/17259

Test plan:
- Build and run `velox_functions_test --gtest_filter=DecimalArithmeticTest.*` in the container
